move CDN methods to CDN class
