+++ classes/Plugin.php
@@ -245,6 +245,7 @@ class Plugin {
 	 */
 	private function get_service_providers() {
 		return [
+			'Imagify\CDN\ServiceProvider',
 			'Imagify\Picture\ServiceProvider',
 			'Imagify\Webp\ServiceProvider',
 		];
